--- example/java/io/orkes/conductor/sdk/examples/DynamicWorkflow/workers/UserInfo.java
@@ -1,52 +1,0 @@
-package io.orkes.conductor.sdk.examples.DynamicWorkflow.workers;
-
-public class UserInfo {
-
-    private String name;
-
-    private String id;
-
-    private String email;
-
-    private String phoneNumber;
-
-    public UserInfo() {}
-
-    public UserInfo(String name, String id) {
-        this.name = name;
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-            this.name = name;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    public void setPhoneNumber(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
-    }
-    }
-
