--- example/java/io/orkes/conductor/sdk/examples/HelloWorld/workflow/WorkflowInput.java
@@ -1,14 +1,0 @@
-package io.orkes.conductor.sdk.examples.HelloWorld.workflow;
-
-public class WorkflowInput {
-    private String name;
-    public WorkflowInput(String name) {
-        this.name = name;
-    }
-    public String getName() {
-        return name;
-    }
-    public void setName(String name) {
-        this.name = name;
-    }
-}
