+++ examples/java/io/orkes/conductor/sdk/DynamicWorkflow/workflow/WorkflowInput.java
@@ -1,0 +1,18 @@
+package io.orkes.conductor.sdk.DynamicWorkflow.workflow;
+
+public class WorkflowInput {
+    private String userId;
+    public WorkflowInput(String userId) {
+        this.userId = userId;
+    }
+
+    public String getUserId() {
+        return userId;
+    }
+
+    public void setUserId(String userId) {
+        this.userId = userId;
+    }    
+}
+
+
