+++ examples/java/io/orkes/conductor/sdk/HelloWorld/workflow/WorkflowInput.java
@@ -1,0 +1,14 @@
+package io.orkes.conductor.sdk.HelloWorld.workflow;
+
+public class WorkflowInput {
+    private String name;
+    public WorkflowInput(String name) {
+        this.name = name;
+    }
+    public String getName() {
+        return name;
+    }
+    public void setName(String name) {
+        this.name = name;
+    }
+}
