+++ examples/java/io/orkes/conductor/sdk/TaskWorkers/pojo/UserInfo.java
@@ -1,0 +1,50 @@
+package io.orkes.conductor.sdk.TaskWorkers.pojo;
+public class UserInfo {
+
+    private String name;
+
+    private String id;
+
+    private String email;
+
+    private String phoneNumber;
+
+    public UserInfo(String orkes) {}
+
+    public UserInfo(String name, String id) {
+        this.name = name;
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+}
